Update Partition size to 16 bits. (#815)
* Update Partition size to 16 bits.

This will allow us to support up to 65k processors, which should be good for the forseeable future.

Needs to be tested a lot, especially if we can access that 256 core system again.

Closes #809

* Fix review and build failures

* Make partition mask calculation a bit cleaner

* Actually save

* Fix a few missed conversion checks

* Fix Cid crash

* Fix a bit of comments

+++ src/core/connection.c
@@ -39,7 +39,7 @@ Abstract:
 typedef struct QUIC_RECEIVE_PROCESSING_STATE {
     BOOLEAN ResetIdleTimeout;
     BOOLEAN UpdatePartitionId;
-    uint8_t PartitionIndex;
+    uint16_t PartitionIndex;
 } QUIC_RECEIVE_PROCESSING_STATE;
 
 _IRQL_requires_max_(PASSIVE_LEVEL)
@@ -66,11 +66,11 @@ QuicConnAlloc(
     // the current processor for now. Once the connection receives a packet the
     // partition can be updated accordingly.
     //
-    uint8_t BasePartitionId =
+    uint16_t BasePartitionId =
         IsServer ?
             (Datagram->PartitionIndex % MsQuicLib.PartitionCount) :
             CurProcIndex % MsQuicLib.PartitionCount;
-    uint8_t PartitionId = QuicPartitionIdCreate(BasePartitionId);
+    uint16_t PartitionId = QuicPartitionIdCreate(BasePartitionId);
     QUIC_DBG_ASSERT(BasePartitionId == QuicPartitionIdGetIndex(PartitionId));
 
     QUIC_CONNECTION* Connection =
@@ -6308,8 +6308,8 @@ QuicConnParamGet(
 
     case QUIC_PARAM_CONN_IDEAL_PROCESSOR:
 
-        if (*BufferLength < sizeof(uint8_t)) {
-            *BufferLength = sizeof(uint8_t);
+        if (*BufferLength < sizeof(uint16_t)) {
+            *BufferLength = sizeof(uint16_t);
             Status = QUIC_STATUS_BUFFER_TOO_SMALL;
             break;
         }
@@ -6319,8 +6319,8 @@ QuicConnParamGet(
             break;
         }
 
-        *BufferLength = sizeof(uint8_t);
-        *(uint8_t*)Buffer = Connection->Worker->IdealProcessor;
+        *BufferLength = sizeof(uint16_t);
+        *(uint16_t*)Buffer = Connection->Worker->IdealProcessor;
 
         Status = QUIC_STATUS_SUCCESS;
         break;
